refactor(archiver): Separate syncing from datastore with ArchiveSourceBase
