+++ yarn-project/archiver/src/archiver/index.ts
@@ -1,6 +1,7 @@
 export * from './archiver.js';
+export * from './archive_source_base.js';
+export * from './archiver_store_updates.js';
 export * from './config.js';
 export { type L1PublishedData } from './structs/published.js';
-export type { ArchiverDataStore } from './archiver_store.js';
 export { KVArchiverDataStore, ARCHIVER_DB_VERSION } from './kv_archiver_store/kv_archiver_store.js';
 export { ContractInstanceStore } from './kv_archiver_store/contract_instance_store.js';
